Refactor and enhance project with new features
- Modularized the codebase by introducing separate header files for each window class.
- Added `Expression` class for mathematical expression evaluation using `exprtk` and corresponding tests.
- Introduced new window classes (`Window_Analysis`, `Window_Console`, `Window_FPS`, `Window_Live`, `Window_OpenProject`, `Window_SignalBrowser`) for improved UI functionality.
- Implemented `ProjectManager` and JSON deserialization for project data.
- Updated `CMakeLists.txt` to include new directories, files, and tests.
- Improved type safety by changing `int` to `size_t` for count-related methods.
- Refactored schema handling with `kindFromString` and `kindToString` utilities.
- Enhanced `storage_test` with additional test cases and updated `StreamOptions`.

+++ include/console.h
@@ -15,8 +15,8 @@ class Console {
 public:
     void log(int prio, std::string text);
     void updateRecentlyAddedItems();
-    int getCount() const;
-    int getCountRecent() const;
+    size_t getCount() const;
+    size_t getCountRecent() const;
     LogItem getItem(int index) const;
     LogItem getItemRecent(int index) const;
     void removeItem(int index);
